+++ editor_patch/vtypes.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace rf
+{
+    struct String
+    {
+        int max_len;
+        char* buf;
+    };
+}